requirements.txt
torch==2.0.0
torchvision==0.15.1
lightning==2.1.3
torchmetrics==1.2.1
numpy==1.23.5
pandas==1.5.3
scipy==1.11.4
anndata==0.8.0
scanpy==1.9.3
squidpy==1.2.2
matplotlib==3.8.2
tqdm==4.66.4
h5py==3.8.0
pyyaml==6.0.1
shapely==2.0.4
pyvips==2.2.3
opencv-python==4.9.0.80
pillow==10.3.0
scikit-learn==1.2.2
numba==0.56.4
llvmlite==0.39.1
dask-image==2022.9.0
huggingface_hub==0.23.4
transformers==4.42.3
timm==0.9.12
plotnine
hest==0.2.0
